api-python/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
openai>=1.40   # new SDK using max_completion_tokens
PyMuPDF>=1.24
jinja2>=3.1
pydantic>=2.5
